New Crystal Desktop Search Applet
This applet is a popupapplet allowing to search through your nepomuk
index. It uses the Nepomuk::SearchHitView to display results. You'll
need nepomuk-visualization/ from playground's nepomuk-kde directory
for this to work.

This applet is meant as a testing ground for concepts around desktop
search, a base to play around with.

By limiting the number of results, we can make the search appear faster,
even fast enough so that "performance is acceptable" on my machine.
while the results look very sensible to me, they're not perfect if you
have the search time out.

svn path=/trunk/playground/base/plasma/applets/crystal/; revision=909460